refactor: consolidated ip address utils
